Added in populate db and install in trial store
Created a pointer at the end of each install to specify that they need to install in the trial store. I think that is where the confusion is coming from. At one point there was an All done! but it really needed to say and install in your trial store.